Implements the resize feature.
Adds a new button to recover the new value that
will be used as the height and width of the new grid.
A new function has been created to remove the existing grid.
The creation of the grid has been moved into a function
to ease the creation of the new grid during the resizing process.